Refactor to typeorm config
