+++ config/typeorm/database.ts
@@ -6,8 +6,8 @@ export const connectionSource = new DataSource({
   host: 'localhost',
   port: 5432,
   username: process.env.DATABASE_USER,
-  password: 'pass',
-  database: 'somehealthchecker',
+  password: 'test',
+  database: process.env.DATABASE_SCHEMA,
   logging: false,
   synchronize: false,
   entities: ['src/**/**.entity{.ts,.js}'],
