+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import configuration from '../config/configuration';
-import typeorm from '../config/typeorm';
+import typeorm from '../config/typeorm/typeorm';
 
 @Module({
   imports: [
